+++ extg4/X4PhysicalVolume.cc
@@ -504,7 +504,7 @@ for persistency convenience.
 void X4PhysicalVolume::convertSolids_r(const G4VPhysicalVolume* const pv, int depth)
 {
     const G4LogicalVolume* lv = pv->GetLogicalVolume() ;
-    for (int i=0 ; i < lv->GetNoDaughters() ;i++ )
+    for (size_t i=0 ; i < lv->GetNoDaughters() ;i++ )
     {
         const G4VPhysicalVolume* const daughter_pv = lv->GetDaughter(i);
         convertSolids_r( daughter_pv , depth + 1 );
@@ -886,7 +886,7 @@ GVolume* X4PhysicalVolume::convertStructure_r(const G4VPhysicalVolume* const pv,
 
      const G4LogicalVolume* const lv = pv->GetLogicalVolume();
   
-     for (int i=0 ; i < lv->GetNoDaughters() ;i++ )
+     for (size_t i=0 ; i < lv->GetNoDaughters() ;i++ )
      {
          const G4VPhysicalVolume* const child_pv = lv->GetDaughter(i);
          convertStructure_r(child_pv, volume, depth+1, pv, recursive_select );
